[Kafka] Kafka management scripts
Two scripts have implemented so as to be allow topic creation and
listing from our machines (the (Vagrant) SSH connection to the cluster
is done within the scripts).

The documentation of the two scripts should be detailed and clear
enough ("-h" option of the scripts). If not, please ask me to update the
related documentation.